Connectivity Issue Between Profile Menu and Account Page Based on Wallet Status #31
A connectivity issue has been identified between the profile menu (ProfileMenu) and the account page (AccountPage) based on the wallet status. In the account page, the wallet connection status is checked, and if not connected, the user is redirected to another page using the redirect("/") function. However, the useWeb3ModalAccount hook is used to check the wallet connection status of the button in the profile menu. Therefore, when the wallet connection status changes (when isConnected changes), the status of the button in the profile menu should be updated, and when the wallet connection status changes, redirection should be made to the account page.

Co-Authored-By: Faru Nuri Sönmez <info@farunurisonmez.com>